+++ ursa_helpdesk/README.rst
@@ -1,0 +1,20 @@
+.. image:: https://img.shields.io/badge/license-AGPLv3-blue.svg
+   :target: https://www.gnu.org/licenses/agpl.html
+   :alt: License: AGPL-3
+
+=============
+Ursa Helpdesk
+=============
+
+Sets reply-to in the helpdesk item. This allows a standard reply-to address
+to be set when chatter is used to respond to help desk issues.
+
+
+Credits
+=======
+
+Contributors
+------------
+
+* Ursa Information Systems <http://www.ursainfosystems.com)>
+* Eficent Business and IT Consulting Services S.L. <contact@eficent.com>
