+++ ursa_helpdesk/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import models
+from . import ursa_helpdesk
+from . import ursa_helpdesk_mail
