+++ ursa_helpdesk/views/ursa_helpdesk.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record id="crm_case_form_view_helpdesk_inherit" model="ir.ui.view">
+        <field name="name">crm.helpdesk.inherit</field>
+        <field name="model">crm.helpdesk</field>
+        <field name="type">form</field>
+        <field name="inherit_id" ref="crm_helpdesk.crm_case_form_view_helpdesk"/>
+        <field name="arch" type="xml">
+            <field name="email_from" position="after">
+                <field name="reply_to" />
+            </field>
+        </field>
+    </record>
+
+    <function model="ir.values" name="set_default"
+              id="default_helpdesk_reply_to"
+              eval="('crm.helpdesk', 'helpdesk_reply_to', 'support@lulzbot.com')"/>
+
+    <function model="ir.values" name="set_default"
+              id="default_replace_helpdesk_email_from"
+              eval="('crm.helpdesk', 'replace_helpdesk_email_from', 1)"/>
+
+</odoo>
